refactor(rebate): rename and restructure rebate templates to use 'rebate-tables'
