+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/page.tsx
@@ -2,5 +2,5 @@ import { redirect } from "next/navigation";
 import { getBaseLink } from "src/utils";
 
 export default function Page() {
-  redirect(getBaseLink("settings/templates/rebate"));
+  redirect(getBaseLink("settings/templates/rebate-tables"));
 }
