+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/rebate-tables/_components/table-data.tsx
@@ -38,7 +38,7 @@ const rebateTableHeadersColumns = (
     },
     links: {
       name: {
-        prefix: `/settings/templates/rebate`,
+        prefix: `/settings/templates/rebate-tables`,
         targetAccessorKey: "id",
       },
     },
