+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/rebate-tables/_components/table.tsx
