+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/rebate-tables/new/page.tsx
@@ -1,5 +1,5 @@
 import { getResourceData } from "src/language-data/unirefund/ContractService";
-import RebateForm from "../rebate-form";
+import RebateForm from "../_components/rebate-form";
 
 export default async function Page({ params }: { params: { lang: string } }) {
   const { languageData } = await getResourceData(params.lang);
