+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/rebate-tables/page.tsx
@@ -2,7 +2,7 @@ import { notFound } from "next/navigation";
 import { getRebateTableHeadersApi } from "src/actions/unirefund/ContractService/action";
 import { isUnauthorized } from "src/utils/page-policy/page-policy";
 import { getResourceData } from "src/language-data/unirefund/ContractService";
-import RebateTable from "./table";
+import RebateTable from "./_components/table";
 
 export default async function Page({
   params,
